refactor(asset): extract asset method
